% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalvalue-funs.R
\name{crit_from_t_t2s}
\alias{crit_from_t_t2s}
\title{crit_from_t_t2s}
\usage{
crit_from_t_t2s(
  t = NULL,
  n1,
  n2,
  se = NULL,
  conf.level,
  hypothesis,
  var.equal = FALSE
)
}
\arguments{
\item{t}{the t value.}

\item{n1}{a number corresponding to the sample size of group 1.}

\item{n2}{a number corresponding to the sample size of group 2.}

\item{se}{a number corresponding to the standard error.}

\item{conf.level}{confidence level of the interval.}

\item{hypothesis}{a character string indicating the alternative hypothesis ("less", "greater" or "two.tailed").}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal.}
}
\value{
the output returns a "d" which is the cohen's d, the critical d which is the minimum value for which to get a significant result with a given sample, the "bc" is the numerator of the formula from which the d is calculated and "df" are the degrees of freedom.
}
\description{
This function allows to calculate the cohen's d and the critical d given the t-value for a two samples t-test, sample size of the two groups and the standard error, specifying the confidence level of the interval, the direction of the hypothesis and the variance parameter.
}
