% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalvalue-funs.R
\name{crit_from_t_t2sp}
\alias{crit_from_t_t2sp}
\title{crit_from_t_t2sp}
\usage{
crit_from_t_t2sp(t = NULL, n, se = NULL, r12 = NULL, hypothesis, conf.level)
}
\arguments{
\item{t}{the t value.}

\item{n}{a number corresponding to the sample size.}

\item{se}{a number corresponding to the standard error.}

\item{r12}{a number corresponding to the correlation between variable 1 and variable 2.}

\item{hypothesis}{a character string indicating the alternative hypothesis ("less", "greater" or "two.tailed").}

\item{conf.level}{confidence level of the interval.}
}
\value{
the output returns a \code{dz} which is the critical Cohen's d standartized on the standard deviation of the differences, the \code{dzc} is the critical standardized d using the pooled standard deviation, the \code{d} is the Cohen's d (pooled standard deviation), the \code{bc} is the numerator of the formula from which the d is calculated, \code{df} are the degrees of freedom, and \code{se} is the standard error.
}
\description{
This function allows to calculate the standardized cohen's d, the critical standardized cohen's d, cohen's d and critical cohen's d given for a paired t test, given the t-value, the sample size, the standard error and the correlation between the two variables, specifying the confidence level of the interval and the direction of the hypothesis.
}
