% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalvalue-funs.R
\name{critical_cor}
\alias{critical_cor}
\title{critical_cor}
\usage{
critical_cor(
  r = NULL,
  n,
  conf.level = 0.95,
  hypothesis = c("two.sided", "greater", "less"),
  test = c("t", "z")
)
}
\arguments{
\item{r}{a number corresponding to the correlation coefficient.}

\item{n}{a number corresponding to the sample size.}

\item{conf.level}{the confidence level to set the confidence interval, default is set to 0.95.}

\item{hypothesis}{a character string indicating the alternative hypothesis ("less", "greater" or "two.tailed").}

\item{test}{a parameter to specify which test to apply, either "t" for a t-test or "z" for a z-test.}
}
\value{
\code{rc} is the critical correlation value, \code{rzc} is the Fisher's z transformed critical correlation, \code{df} are the degrees of freedom, \code{se_r} is the standard error of the observed correlation, \code{se_rc} is the standard error of the critical correlation, \code{se_rzc} is the standard error of the Fisher's z transformed critical correlation and \code{test} is the statistical test (either t or z).
}
\description{
This function allows to calculate the critical correlation value. When using \code{test = "t"}, the critical value is calculated assuming a Student t distribution with \eqn{n - 2} degrees of freedom and the standard error calculated using the raw correlation coefficient. When \code{test = "z"}, the critical value is calculated assimong a standard Normal distribution and the standard error is calculated applying the Fisher's z transformation.
}
