% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalvalue-funs.R
\name{critical_t2sp}
\alias{critical_t2sp}
\title{critical_t2sp}
\usage{
critical_t2sp(
  m1 = NULL,
  m2 = NULL,
  t = NULL,
  sd1 = NULL,
  sd2 = NULL,
  r12 = NULL,
  n,
  se = NULL,
  hypothesis = c("two.sided", "greater", "less"),
  conf.level = 0.95
)
}
\arguments{
\item{m1}{a number representing the mean of group 1.}

\item{m2}{a number representing the mean of group 2.}

\item{t}{the t value.}

\item{sd1}{a number representing the standard deviation of group 1.}

\item{sd2}{a number representing the standard deviation of group 2.}

\item{r12}{a number corresponding to the correlation between variable 1 and variable 2.}

\item{n}{a number corresponding to the sample size.}

\item{se}{a number corresponding to the standard error.}

\item{hypothesis}{a character string indicating the alternative hypothesis ("less", "greater" or "two.tailed").}

\item{conf.level}{the confidence level to set the confidence interval, default is set to 0.95.}
}
\value{
the output returns a \code{dz} which is the critical Cohen's d standartized on the standard deviation of the differences, the \code{dzc} is the critical standardized d using the pooled standard deviation, the \code{d} is the Cohen's d, the \code{dc} is the minimum value for which to get a significant result with a given sample, the \code{bc} is the numerator of the formula from which the d is calculated, \code{se} is the standard error, \code{df} are the degrees of freedom,the \code{g} and \code{gc} are respectively \code{d} and \code{dc} with Hedges's Correction for small samples and \code{gz} and \code{gzc} are the standardized ones.
}
\description{
The function allows to calculate the standardized cohen's d, the critical standardized cohen's d, the cohen's d and the critical d for a paired two samples t-test.
}
\examples{
# critical value from summary statistics
m1 <- 10
m2 <- 15
sd1 <- 5
sd2 <- 4.25
n <- 30
critical_t2sp(m1 = m1, m2 = m2, sd1 = sd1, sd2 = sd2, n = n)
# critical value from the t statistic
se <- sqrt((sd1^2 + sd2^2) / n)
t <- (m1 - m2) / se
critical_t2sp(t = t, n = n, se = se) # se only required for calculating bc
}
