% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalvalue-funs.R
\name{critical_t1s}
\alias{critical_t1s}
\title{critical_t1s}
\usage{
critical_t1s(
  m = NULL,
  s = NULL,
  t = NULL,
  n,
  se = NULL,
  hypothesis = c("two.sided", "greater", "less"),
  conf.level = 0.95
)
}
\arguments{
\item{m}{a number representing the mean of the group.}

\item{s}{Variance}

\item{t}{the t value.}

\item{n}{a number corresponding to the sample size.}

\item{se}{a number corresponding to the standard error.}

\item{hypothesis}{a character string indicating the alternative hypothesis ("less", "greater" or "two.tailed").}

\item{conf.level}{the confidence level to set the confidence interval, default is set to 0.95.}
}
\value{
the output returns, based on the inserted values, mean, variance, confidence interval and hypothesis or t-value, numerosity, standard error, confidence interval and hypothesis.
}
\description{
The function allows to calculate cohen's d and critical d for a one-sample t-test.
}
\examples{
# critical value from summary statistics
m <- 0.5
s <- 1
n <- 30
critical_t1s(m = m, s = s, n = n)
# critical value from the t statistic
se <- s / sqrt(n)
t <- m / se
critical_t1s(t = t, n = n, se = se) # se only required for calculating bc
}
